Added monitor. (Not working fine)

+++ philo/includes/utils.h
@@ -6,7 +6,7 @@
 /*   By: muayna <muayna@student.42istanbul.com.t    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/26 16:04:24 by muayna            #+#    #+#             */
-/*   Updated: 2026/04/26 21:52:04 by muayna           ###   ########.fr       */
+/*   Updated: 2026/06/06 12:42:13 by muayna           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include "philo.h"
 # include "../libs/garbage_collector/gb.h"
 
-long    ft_atoi(const char *nptr);
+long long   ft_atoi(const char *nptr);
 int     ft_isdigit(int c);
 size_t  ft_strlen(const char *s);
